+++ packages/core/src/hooks/useContractCall.ts
@@ -2,7 +2,10 @@ import { Interface } from '@ethersproject/abi'
 import { useMemo } from 'react'
 import { Falsy } from '../model/types'
 import { useChainCalls } from './useChainCalls'
-import { ChainCall } from '../providers/chainState/callsReducer'
+import { RawCall, useNetwork } from '../providers'
+import { ChainId } from '../constants'
+import { QueryParams } from '../constants/type/QueryParams'
+import type { useCall, useCalls } from './useCall'
 
 function warnOnInvalidContractCall(call: ContractCall | Falsy) {
   console.warn(
@@ -10,7 +13,7 @@ function warnOnInvalidContractCall(call: ContractCall | Falsy) {
   )
 }
 
-function encodeCallData(call: ContractCall | Falsy): ChainCall | Falsy {
+function encodeCallData(call: ContractCall | Falsy, chainId: ChainId): RawCall | Falsy {
   if (!call) {
     return undefined
   }
@@ -19,13 +22,17 @@ function encodeCallData(call: ContractCall | Falsy): ChainCall | Falsy {
     return undefined
   }
   try {
-    return { address: call.address, data: call.abi.encodeFunctionData(call.method, call.args) }
+    return { address: call.address, data: call.abi.encodeFunctionData(call.method, call.args), chainId }
   } catch {
     warnOnInvalidContractCall(call)
     return undefined
   }
 }
 
+/**
+ * @public
+ * @deprecated Use {@link useCall} instead.
+ */
 export interface ContractCall {
   abi: Interface
   address: string
@@ -33,12 +40,28 @@ export interface ContractCall {
   args: any[]
 }
 
-export function useContractCall(call: ContractCall | Falsy): any[] | undefined {
-  return useContractCalls([call])[0]
+/**
+ * @public
+ * @deprecated Use {@link useCalls} instead.
+ */
+export function useContractCall(call: ContractCall | Falsy, queryParams: QueryParams = {}): any[] | undefined {
+  return useContractCalls([call], queryParams)[0]
 }
 
-export function useContractCalls(calls: (ContractCall | Falsy)[]): (any[] | undefined)[] {
-  const results = useChainCalls(calls.map(encodeCallData))
+/**
+ * @public
+ * @deprecated Use {@link useCall} instead.
+ */
+export function useContractCalls(
+  calls: (ContractCall | Falsy)[],
+  queryParams: QueryParams = {}
+): (any[] | undefined)[] {
+  const { network } = useNetwork()
+  const chainId = queryParams.chainId ?? network.chainId
+
+  const results = useChainCalls(
+    calls.map((call) => (chainId !== undefined ? encodeCallData(call, chainId) : undefined))
+  )
 
   return useMemo(
     () =>