+++ packages/core/src/hooks/useTokenList.ts
@@ -1,15 +1,20 @@
 import { useEffect, useState } from 'react'
 import { TokenInfo } from '@uniswap/token-lists'
-import { ChainId } from '../constants'
 import { useEthers } from './useEthers'
 
-interface TokenList {
+/**
+ * @public
+ */
+export interface TokenList {
   name: string
   logoURI: string
   tokens: TokenInfo[]
 }
 
-export function useTokenList(tokenListURI: string, overrideChainId?: ChainId, tags?: string[]) {
+/**
+ * @public
+ */
+export function useTokenList(tokenListURI: string, overrideChainId?: number, tags?: string[]): TokenList | undefined {
   const { chainId: providerChainId } = useEthers()
   const [tokenList, setTokenList] = useState<TokenList>()