+++ packages/core/src/providers/blockNumber/blockNumber/context.ts
@@ -1,0 +1,10 @@
+import { createContext, useContext } from 'react'
+
+export const BlockNumberContext = createContext<number | undefined>(undefined)
+
+/**
+ * @public
+ */
+export function useBlockNumber(): number | undefined {
+  return useContext(BlockNumberContext)
+}