+++ packages/core/src/providers/network/readonlyNetworks/model.ts
@@ -1,0 +1,6 @@
+import { ChainId } from '../../../constants'
+import { JsonRpcProvider } from '@ethersproject/providers'
+
+export type Providers = {
+  [chainId in ChainId]?: JsonRpcProvider
+}