+++ packages/core/src/testing/utils/deployMulticall.ts
@@ -1,0 +1,12 @@
+import { deployContract, MockProvider } from 'ethereum-waffle'
+import { MultiCall } from '../../constants'
+
+export const deployMulticall = async (provider: MockProvider, chainId: number) => {
+  const multicall = await deployContract((await provider.getWallets())[0], {
+    bytecode: MultiCall.bytecode,
+    abi: MultiCall.abi,
+  })
+  const multicallAddresses = { [chainId]: multicall.address }
+
+  return multicallAddresses
+}