+++ packages/extension/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @usedapp/extension
 
+## 0.3.1
+
+### Patch Changes
+
+- 378177c: Fetching state from multiple chains simultaneously
+- ad021b0: Use pnpm instead of yarn
+
+## 0.3.0
+
+### Minor Changes
+
+- fd8e924: Remove @web3-react dependency, introduce own way of provider management
+
 ## 0.2.0
 
 ### Minor Changes